refactor: sshnp to spec/impl and SSHNPParams class
