+++ packages/sshnoports/lib/sshnp/sshnp.dart
@@ -5,57 +5,47 @@ import 'dart:io';
 // atPlatform packages
 import 'package:at_utils/at_logger.dart';
 import 'package:at_client/at_client.dart';
-import 'package:at_onboarding_cli/at_onboarding_cli.dart';
 
 // other packages
-import 'package:args/args.dart';
-import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
-import 'package:sshnoports/sshnp/parse_config_file.dart';
-import 'package:sshnoports/sshnp/sshnp_arg.dart';
-import 'package:uuid/uuid.dart';
-import 'package:version/version.dart';
+import 'package:sshnoports/sshnp/sshnp_impl.dart';
 
 // local packages
-import 'package:sshnoports/common/service_factories.dart';
-import 'package:sshnoports/common/utils.dart';
-import 'package:sshnoports/sshnp/cleanup.dart';
-import 'package:sshnoports/version.dart';
 
-class SSHNP {
+abstract class SSHNP {
   // TODO Make this a const in SSHRVD class
   static const String sshrvdNameSpace = 'sshrvd';
 
-  final AtSignLogger logger = AtSignLogger(' sshnp ');
+  abstract final AtSignLogger logger;
 
   // ====================================================================
   // Final instance variables, injected via constructor
   // ====================================================================
   /// The [AtClient] used to communicate with sshnpd and sshrvd
-  final AtClient atClient;
+  abstract final AtClient atClient;
 
   /// The atSign of the sshnpd we wish to communicate with
-  final String sshnpdAtSign;
+  abstract final String sshnpdAtSign;
 
   /// The device name of the sshnpd we wish to communicate with
-  final String device;
+  abstract final String device;
 
   /// The user name on this host
-  final String username;
+  abstract final String username;
 
   /// The home directory on this host
-  final String homeDirectory;
+  abstract final String homeDirectory;
 
   /// The sessionId we will use
-  final String sessionId;
+  abstract final String sessionId;
 
-  final String sendSshPublicKey;
-  final List<String> localSshOptions;
+  abstract final String sendSshPublicKey;
+  abstract final List<String> localSshOptions;
 
   /// When false, we generate [sshPublicKey] and [sshPrivateKey] using ed25519.
   /// When true, we generate [sshPublicKey] and [sshPrivateKey] using RSA.
   /// Defaults to false
-  final bool rsa;
+  abstract final bool rsa;
 
   // ====================================================================
   // Volatile instance variables, injected via constructor
@@ -65,17 +55,17 @@ class SSHNP {
   /// Host that we will send to sshnpd for it to connect to,
   /// or the atSign of the sshrvd.
   /// If using sshrvd then we will fetch the _actual_ host to use from sshrvd.
-  String host;
+  abstract String host;
 
   /// Port that we will send to sshnpd for it to connect to.
   /// Required if we are not using sshrvd.
   /// If using sshrvd then initial port value will be ignored and instead we
   /// will fetch the port from sshrvd.
-  String port;
+  abstract String port;
 
   /// Port to which sshnpd will forwardRemote its [SSHClient]. If localPort
   /// is set to '0' then
-  String localPort;
+  abstract String localPort;
 
   // ====================================================================
   // Derived final instance variables, set during construction or init
@@ -83,49 +73,49 @@ class SSHNP {
 
   /// Set to [AtClient.getCurrentAtSign] during construction
   @visibleForTesting
-  late final String clientAtSign;
+  abstract final String clientAtSign;
 
   /// The username to use on the remote host in the ssh session. Either passed
   /// through class constructor or fetched from the sshnpd
   /// by [fetchRemoteUserName] during [init]
-  String? remoteUsername;
+  abstract String? remoteUsername;
 
   /// Set by [generateSshKeys] during [init].
   /// sshnp generates a new keypair for each ssh session, using ed25519 by
   /// default but rsa if the [rsa] flag is set to true. sshnp will write
   /// [sshPublicKey] to ~/.ssh/authorized_keys
-  late final String sshPublicKey;
+  abstract final String sshPublicKey;
 
   /// Set by [generateSshKeys] during [init].
   /// sshnp generates a new keypair for each ssh session, using ed25519 by
   /// default but rsa if the [rsa] flag is set to true. sshnp will send the
   /// [sshPrivateKey] to sshnpd
-  late final String sshPrivateKey;
+  abstract final String sshPrivateKey;
 
   /// Namespace will be set to [device].sshnp
-  late final String nameSpace;
+  abstract final String nameSpace;
 
   /// When using sshrvd, this is fetched from sshrvd during [init]
-  late final String sshrvdPort;
+  abstract final String sshrvdPort;
 
   /// Set to '$localPort $port $username $host $sessionId' during [init]
-  late final String sshString;
+  abstract final String sshString;
 
   /// Set by constructor to
   /// '$homeDirectory${Platform.pathSeparator}.ssh${Platform.pathSeparator}'
-  late final String sshHomeDirectory;
+  abstract final String sshHomeDirectory;
 
   /// true once we have received any response (success or error) from sshnpd
   @visibleForTesting
-  bool sshnpdAck = false;
+  abstract bool sshnpdAck;
 
   /// true once we have received an error response from sshnpd
   @visibleForTesting
-  bool sshnpdAckErrors = false;
+  abstract bool sshnpdAckErrors;
 
   /// true once we have received a response from sshrvd
   @visibleForTesting
-  bool sshrvdAck = false;
+  abstract bool sshrvdAck;
 
   // In the future (perhaps) we can send other commands
   // Perhaps OpenVPN or shell commands
@@ -135,32 +125,42 @@ class SSHNP {
   @visibleForTesting
   bool initialized = false;
 
-  SSHNP({
+  factory SSHNP({
     // final fields
-    required this.atClient,
-    required this.sshnpdAtSign,
-    required this.device,
-    required this.username,
-    required this.homeDirectory,
-    required this.sessionId,
-    this.sendSshPublicKey = 'false',
-    required this.localSshOptions,
-    this.rsa = false,
+    required AtClient atClient,
+    required String sshnpdAtSign,
+    required String device,
+    required String username,
+    required String homeDirectory,
+    required String sessionId,
+    String sendSshPublicKey = 'false',
+    required List<String> localSshOptions,
+    bool rsa = false,
     // volatile fields
-    required this.host,
-    required this.port,
-    required this.localPort,
-    this.remoteUsername,
+    required String host,
+    required String port,
+    required String localPort,
+    String? remoteUsername,
   }) {
-    nameSpace = '$device.sshnp';
-    clientAtSign = atClient.getCurrentAtSign()!;
-    logger.hierarchicalLoggingEnabled = true;
-    logger.logger.level = Level.SHOUT;
+    return SSHNPImpl(
+      atClient: atClient,
+      sshnpdAtSign: sshnpdAtSign,
+      device: device,
+      username: username,
+      homeDirectory: homeDirectory,
+      sessionId: sessionId,
+      sendSshPublicKey: sendSshPublicKey,
+      localSshOptions: localSshOptions,
+      rsa: rsa,
+      host: host,
+      port: port,
+      localPort: localPort,
+      remoteUsername: remoteUsername,
+    );
+  }
 
-    sshHomeDirectory = getDefaultSshDirectory(homeDirectory);
-    if (!Directory(sshHomeDirectory).existsSync()) {
-      Directory(sshHomeDirectory).createSync();
-    }
+  static Future<SSHNP> fromCommandLineArgs(List<String> args) async {
+    return SSHNPImpl.fromCommandLineArgs(args);
   }
 
   /// Must be run after construction, to complete initialization
@@ -174,120 +174,14 @@ class SSHNP {
   ///   from sshrvd
   /// - calls [sharePrivateKeyWithSshnpd]
   /// - calls [sharePublicKeyWithSshnpdIfRequired]
-  Future<void> init() async {
-    if (initialized) {
-      throw StateError('Cannot init() - already initialized');
-    }
-
-    if (!(await atSignIsActivated(atClient, sshnpdAtSign))) {
-      throw ('sshnpd atSign $sshnpdAtSign is not activated.');
-    }
-
-    logger.info('Subscribing to notifications on $sessionId.$nameSpace@');
-    // Start listening for response notifications from sshnpd
-    atClient.notificationService
-        .subscribe(regex: '$sessionId.$nameSpace@', shouldDecrypt: true)
-        .listen(handleSshnpdResponses);
-
-    await generateSshKeys();
-
-    if (remoteUsername == null) {
-      await fetchRemoteUserName();
-    }
-
-    // find a spare local port
-    if (localPort == '0') {
-      ServerSocket serverSocket =
-          await ServerSocket.bind(InternetAddress.loopbackIPv4, 0);
-      localPort = serverSocket.port.toString();
-      await serverSocket.close();
-    }
-
-    // If host has an @ then contact the sshrvd service for some ports
-    if (host.startsWith('@')) {
-      await getHostAndPortFromSshrvd();
-    }
-
-    if (commandToSend == 'sshd') {
-      // Local port, port of sshd , username , hostname
-      sshString = '$localPort $port $username $host $sessionId';
-    }
-
-    await sharePrivateKeyWithSshnpd();
-
-    await sharePublicKeyWithSshnpdIfRequired();
-
-    initialized = true;
-  }
+  Future<void> init();
 
   /// May only be run after [init] has been run.
   /// - Sends request to sshnpd; the response listener was started by [init]
   /// - Waits for success or error response, or time out after 10 secs
   /// - If got a success response, print the ssh command to use to stdout
   /// - Clean up temporary files
-  Future<void> run() async {
-    if (!initialized) {
-      throw StateError('Cannot run() - not initialized');
-    }
-    AtKey keyForCommandToSend = AtKey()
-      ..key = commandToSend
-      ..namespace = nameSpace
-      ..sharedBy = clientAtSign
-      ..sharedWith = sshnpdAtSign
-      ..metadata = (Metadata()
-        ..ttr = -1
-        ..ttl = 10000);
-
-    try {
-      await atClient.notificationService.notify(
-          NotificationParams.forUpdate(keyForCommandToSend, value: sshString),
-          onSuccess: (notification) {
-        logger.info('SUCCESS:$notification $sshString');
-      }, onError: (notification) {
-        logger.info('ERROR:$notification $sshString');
-      });
-    } catch (e) {
-      stderr.writeln(e.toString());
-      rethrow;
-    }
-
-    // Before we clean up we need to make sure that the reverse ssh made the connection.
-    // Or that if it had a problem what the problem was, or timeout and explain why.
-
-    int counter = 0;
-    // Timer to timeout after 10 Secs or after the Ack of connected/Errors
-    while (!sshnpdAck) {
-      await Future.delayed(Duration(milliseconds: 100));
-      counter++;
-      if (counter == 100) {
-        await cleanUp(sessionId, logger);
-        stderr.writeln('sshnp: connection timeout');
-        exit(1);
-      }
-    }
-
-    // Clean Up the files we created
-    await cleanUp(sessionId, logger);
-
-    // print out base ssh command if we hit no Ack Errors
-    // If we had a Public key include the private key in the command line
-    // By removing the .pub extn
-    if (!sshnpdAckErrors) {
-      if (sendSshPublicKey != 'false') {
-        stdout.write(
-            "ssh -p $localPort $remoteUsername@localhost -i ${sendSshPublicKey.replaceFirst(RegExp(r'.pub$'), '')} ");
-      } else {
-        stdout.write("ssh -p $localPort $remoteUsername@localhost ");
-      }
-      // print out optional arguments
-      for (var argument in localSshOptions) {
-        stdout.write("$argument ");
-      }
-    }
-    // Print the  return
-    stdout.write('\n');
-    exit(0);
-  }
+  Future<void> run();
 
   /// Function which the response subscription (created in the [init] method
   /// will call when it gets a response from the sshnpd
@@ -315,358 +209,15 @@ class SSHNP {
   /// sshnpd, then we expect a key to have been shared whose ID is
   /// @human:username.device.sshnp@daemon
   /// Is not called if remoteUserName was set via constructor
-  Future<void> fetchRemoteUserName() async {
-    AtKey userNameRecordID =
-        AtKey.fromString('$clientAtSign:username.$nameSpace$sshnpdAtSign');
-    try {
-      remoteUsername = (await atClient.get(userNameRecordID)).value as String;
-    } catch (e) {
-      stderr.writeln("Device \"$device\" unknown, or username not shared ");
-      await cleanUp(sessionId, logger);
-      exit(1);
-    }
-  }
+  Future<void> fetchRemoteUserName();
 
-  Future<void> sharePublicKeyWithSshnpdIfRequired() async {
-    if (sendSshPublicKey != 'false') {
-      try {
-        String toSshPublicKey = await File(sendSshPublicKey).readAsString();
-        if (!toSshPublicKey.startsWith('ssh-')) {
-          throw ('$sshHomeDirectory$sendSshPublicKey does not look like a public key file');
-        }
-        AtKey sendOurPublicKeyToSshnpd = AtKey()
-          ..key = 'sshpublickey'
-          ..sharedBy = clientAtSign
-          ..sharedWith = sshnpdAtSign
-          ..metadata = (Metadata()
-            ..ttr = -1
-            ..ttl = 10000);
-        await atClient.notificationService.notify(
-            NotificationParams.forUpdate(sendOurPublicKeyToSshnpd,
-                value: toSshPublicKey), onSuccess: (notification) {
-          logger.info('SUCCESS:$notification');
-        }, onError: (notification) {
-          logger.info('ERROR:$notification');
-        });
-      } catch (e) {
-        stderr.writeln(
-            "Error opening or validating public key file or sending to remote atSign: $e");
-        await cleanUp(sessionId, logger);
-        exit(1);
-      }
-    }
-  }
+  Future<void> sharePublicKeyWithSshnpdIfRequired();
 
-  Future<void> sharePrivateKeyWithSshnpd() async {
-    AtKey sendOurPrivateKeyToSshnpd = AtKey()
-      ..key = 'privatekey'
-      ..sharedBy = clientAtSign
-      ..sharedWith = sshnpdAtSign
-      ..namespace = nameSpace
-      ..metadata = (Metadata()
-        ..ttr = -1
-        ..ttl = 10000);
+  Future<void> sharePrivateKeyWithSshnpd();
 
-    try {
-      await atClient.notificationService.notify(
-          NotificationParams.forUpdate(sendOurPrivateKeyToSshnpd,
-              value: sshPrivateKey), onSuccess: (notification) {
-        logger.info('SUCCESS:$notification');
-      }, onError: (notification) {
-        logger.info('ERROR:$notification');
-      });
-    } catch (e) {
-      stderr.writeln(e.toString());
-      rethrow;
-    }
-  }
+  Future<void> getHostAndPortFromSshrvd();
 
-  Future<void> getHostAndPortFromSshrvd() async {
-    atClient.notificationService
-        .subscribe(regex: '$sessionId.$sshrvdNameSpace@', shouldDecrypt: true)
-        .listen((notification) async {
-      String ipPorts = notification.value.toString();
-      List results = ipPorts.split(',');
-      host = results[0];
-      port = results[1];
-      sshrvdPort = results[2];
-      sshrvdAck = true;
-    });
-
-    AtKey ourSshrvdIdKey = AtKey()
-      ..key = '$device.$sshrvdNameSpace'
-      ..sharedBy = clientAtSign // shared by us
-      ..sharedWith = host // shared with the sshrvd host
-      ..metadata = (Metadata()
-        // as we are sending a notification to the sshrvd namespace,
-        // we don't want to append our namespace
-        ..namespaceAware = false
-        ..ttr = -1
-        ..ttl = 10000);
-
-    try {
-      await atClient.notificationService.notify(
-          NotificationParams.forUpdate(ourSshrvdIdKey, value: sessionId),
-          onSuccess: (notification) {
-        logger.info('SUCCESS:$notification $ourSshrvdIdKey');
-      }, onError: (notification) {
-        logger.info('ERROR:$notification $ourSshrvdIdKey');
-      });
-    } catch (e) {
-      stderr.writeln(e.toString());
-      rethrow;
-    }
-
-    int counter = 0;
-    while (!sshrvdAck) {
-      await Future.delayed(Duration(milliseconds: 100));
-      counter++;
-      if (counter == 100) {
-        await cleanUp(sessionId, logger);
-        stderr.writeln('sshnp: connection timeout to sshrvd $host service');
-        exit(1);
-      }
-    }
-
-    // Connect to rendezvous point using background process.
-    // sshnp (this program) can then exit without issue.
-    unawaited(Process.run(getSshrvCommand(), [host, sshrvdPort]));
-  }
-
-  Future<void> generateSshKeys() async {
-    if (rsa) {
-      await Process.run(
-          'ssh-keygen',
-          [
-            '-t',
-            'rsa',
-            '-b',
-            '4096',
-            '-f',
-            '${sessionId}_sshnp',
-            '-q',
-            '-N',
-            ''
-          ],
-          workingDirectory: sshHomeDirectory);
-    } else {
-      await Process.run(
-          'ssh-keygen',
-          [
-            '-t',
-            'ed25519',
-            '-a',
-            '100',
-            '-f',
-            '${sessionId}_sshnp',
-            '-q',
-            '-N',
-            ''
-          ],
-          workingDirectory: sshHomeDirectory);
-    }
-
-    sshPublicKey =
-        await File('$sshHomeDirectory${sessionId}_sshnp.pub').readAsString();
-    sshPrivateKey =
-        await File('$sshHomeDirectory${sessionId}_sshnp').readAsString();
-
-    // Set up a safe authorized_keys file, for the reverse ssh tunnel
-    File('${sshHomeDirectory}authorized_keys').writeAsStringSync(
-        'command="echo \\"ssh session complete\\";sleep 20",PermitOpen="localhost:22" ${sshPublicKey.trim()} $sessionId\n',
-        mode: FileMode.append);
-  }
-
-  static SSHNPParams parseSSHNPParams(List<String> args) {
-    var p = SSHNPParams();
-    ArgParser parser = createArgParser();
-    late ArgResults r;
-    List<String> configArgs = [];
-
-    // Config file
-    r = parser.parse(args);
-    if (r.wasParsed('config-file')) {
-      configArgs = parseConfigFile(r['config-file']);
-    }
-
-    // Main Args
-    r = parser.parse(configArgs + args);
-
-    // Do we have a username ?
-    p.username = getUserName(throwIfNull: true)!;
-
-    // Do we have a 'home' directory?
-    p.homeDirectory = getHomeDirectory(throwIfNull: true)!;
-
-    p.clientAtSign = r['from'];
-    p.sshnpdAtSign = r['to'];
-
-    // Find atSign key file
-    if (r['key-file'] != null) {
-      p.atKeysFilePath = r['key-file'];
-    } else {
-      p.atKeysFilePath =
-          getDefaultAtKeysFilePath(p.homeDirectory, p.clientAtSign);
-    }
-
-    // Check device string only contains ascii
-    if (checkNonAscii(r['device'])) {
-      throw ('\nDevice name can only contain alphanumeric characters with a max length of 15');
-    }
-
-    p.device = r['device'];
-
-    // Check the public key if the option was selected
-    var sendSshPublicKey = r['ssh-public-key'];
-    if ((sendSshPublicKey != 'false')) {
-      sendSshPublicKey =
-          '${getDefaultSshDirectory(p.homeDirectory)}$sendSshPublicKey';
-      if (!sendSshPublicKey.endsWith('.pub')) {
-        throw ('\n The ssh public key should have a ".pub" extension');
-      }
-    }
-    p.sendSshPublicKey = sendSshPublicKey;
-
-    p.host = r['host'];
-    p.port = r['port'];
-    p.localPort = r['local-port'];
-
-    p.localSshOptions = r['local-ssh-options'];
-
-    p.rsa = r['rsa'];
-    p.verbose = r['verbose'];
-
-    p.remoteUsername = r['remote-user-name'];
-
-    return p;
-  }
-
-  static Future<SSHNP> fromCommandLineArgs(List<String> args) async {
-    try {
-      var p = parseSSHNPParams(args);
-
-      // Check atKeyFile selected exists
-      if (!File(p.atKeysFilePath).existsSync()) {
-        throw ('\n Unable to find .atKeys file : ${p.atKeysFilePath}');
-      }
-
-      if (p.sendSshPublicKey != 'false') {
-        if (!File(p.sendSshPublicKey).existsSync()) {
-          throw ('\n Unable to find ssh public key file : ${p.sendSshPublicKey}');
-        }
-      }
-
-      String sessionId = Uuid().v4();
-
-      AtSignLogger.root_level = 'SHOUT';
-      if (p.verbose) {
-        AtSignLogger.root_level = 'INFO';
-      }
-
-      AtClient atClient = await createAtClient(
-          clientAtSign: p.clientAtSign,
-          device: p.device,
-          sessionId: sessionId,
-          atKeysFilePath: p.atKeysFilePath);
-
-      var sshnp = SSHNP(
-        atClient: atClient,
-        sshnpdAtSign: p.sshnpdAtSign,
-        username: p.username,
-        homeDirectory: p.homeDirectory,
-        sessionId: sessionId,
-        device: p.device,
-        host: p.host,
-        port: p.port,
-        localPort: p.localPort,
-        localSshOptions: p.localSshOptions,
-        rsa: p.rsa,
-        sendSshPublicKey: p.sendSshPublicKey,
-        remoteUsername: p.remoteUsername,
-      );
-      if (p.verbose) {
-        sshnp.logger.logger.level = Level.INFO;
-      }
-
-      return sshnp;
-    } catch (e) {
-      version();
-      stdout.writeln(createArgParser().usage);
-      stderr.writeln(e);
-      exit(1);
-    }
-  }
-
-  static Future<AtClient> createAtClient(
-      {required String clientAtSign,
-      required String device,
-      required String sessionId,
-      required String atKeysFilePath}) async {
-    // Now on to the atPlatform startup
-    //onboarding preference builder can be used to set onboardingService parameters
-    AtOnboardingPreference atOnboardingConfig = AtOnboardingPreference()
-      ..hiveStoragePath = '/tmp/.sshnp/$clientAtSign/$sessionId/storage'
-          .replaceAll('/', Platform.pathSeparator)
-      ..namespace = '$device.sshnp'
-      ..downloadPath =
-          '/tmp/.sshnp/files'.replaceAll('/', Platform.pathSeparator)
-      ..isLocalStoreRequired = true
-      ..commitLogPath = '/tmp/.sshnp/$clientAtSign/$sessionId/storage/commitLog'
-          .replaceAll('/', Platform.pathSeparator)
-      ..fetchOfflineNotifications = false
-      ..atKeysFilePath = atKeysFilePath
-      ..atProtocolEmitted = Version(2, 0, 0);
-
-    AtOnboardingService onboardingService = AtOnboardingServiceImpl(
-        clientAtSign, atOnboardingConfig,
-        atServiceFactory: ServiceFactoryWithNoOpSyncService());
-
-    await onboardingService.authenticate();
-
-    return AtClientManager.getInstance().atClient;
-  }
-
-  static ArgParser createArgParser({bool withConfig = true}) {
-    var parser = ArgParser();
-    // Basic arguments
-    for (SSHNPArg arg in sshnpArgs) {
-      switch (arg.format) {
-        case ArgFormat.option:
-          parser.addOption(
-            arg.name,
-            abbr: arg.abbr,
-            mandatory: arg.mandatory,
-            defaultsTo: arg.defaultsTo as String?,
-            help: arg.help,
-          );
-          break;
-        case ArgFormat.multiOption:
-          parser.addMultiOption(
-            arg.name,
-            abbr: arg.abbr,
-            defaultsTo: arg.defaultsTo as List<String>?,
-            help: arg.help,
-          );
-          break;
-        case ArgFormat.flag:
-          parser.addFlag(
-            arg.name,
-            abbr: arg.abbr,
-            defaultsTo: arg.defaultsTo as bool?,
-            help: arg.help,
-          );
-          break;
-      }
-    }
-    if (withConfig) {
-      parser.addOption(
-        'config-file',
-        help:
-            'Read args from a config file\nMandatory args are not required if already supplied in the config file',
-      );
-    }
-    return parser;
-  }
+  Future<void> generateSshKeys();
 
   /// Return the command which this program should execute in order to start the
   /// sshrv program.
@@ -687,21 +238,4 @@ class SSHNP {
           'sshnp is expected to be run as a compiled executable, not via the dart command');
     }
   }
-}
-
-class SSHNPParams {
-  late final String clientAtSign;
-  late final String sshnpdAtSign;
-  late final String device;
-  late final String host;
-  late final String port;
-  late final String localPort;
-  late final String username;
-  late final String homeDirectory;
-  late final String atKeysFilePath;
-  late final String sendSshPublicKey;
-  late final List<String> localSshOptions;
-  late final bool rsa;
-  late final bool verbose;
-  late final String? remoteUsername;
 }
