+++ packages/sshnoports/test/sshnp_test.dart
@@ -1,12 +1,12 @@
 import 'package:args/args.dart';
-import 'package:sshnoports/sshnp/sshnp.dart';
 import 'package:sshnoports/common/utils.dart';
+import 'package:sshnoports/sshnp/sshnp_params.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('args parser tests', () {
     test('test mandatory args', () {
-      ArgParser parser = SSHNP.createArgParser();
+      ArgParser parser = SSHNPPartialParams.parser;
       // As of version 2.4.2 of the args package, exceptions regarding
       // mandatory options are not thrown when the args are parsed,
       // but when trying to retrieve a mandatory option.
@@ -15,16 +15,16 @@ void main() {
       List<String> args = [];
       expect(() => parser.parse(args)['from'], throwsA(isA<ArgumentError>()));
 
-      args.addAll(['-f','@alice']);
+      args.addAll(['-f', '@alice']);
       expect(parser.parse(args)['from'], '@alice');
       expect(() => parser.parse(args)['to'], throwsA(isA<ArgumentError>()));
 
-      args.addAll(['-t','@bob']);
+      args.addAll(['-t', '@bob']);
       expect(parser.parse(args)['from'], '@alice');
       expect(parser.parse(args)['to'], '@bob');
       expect(() => parser.parse(args)['host'], throwsA(isA<ArgumentError>()));
 
-      args.addAll(['-h','host.subdomain.test']);
+      args.addAll(['-h', 'host.subdomain.test']);
       expect(parser.parse(args)['from'], '@alice');
       expect(parser.parse(args)['to'], '@bob');
       expect(parser.parse(args)['host'], 'host.subdomain.test');
@@ -35,7 +35,7 @@ void main() {
       args.addAll(['-f', '@alice']);
       args.addAll(['-t', '@bob']);
       args.addAll(['-h', 'host.subdomain.test']);
-      var p = SSHNP.parseSSHNPParams(args);
+      var p = SSHNPParams.fromPartial(SSHNPPartialParams.fromArgs(args));
       expect(p.clientAtSign, '@alice');
       expect(p.sshnpdAtSign, '@bob');
       expect(p.host, 'host.subdomain.test');
@@ -44,8 +44,9 @@ void main() {
       expect(p.port, '22');
       expect(p.localPort, '0');
       expect(p.username, getUserName(throwIfNull: true));
-      expect(p.homeDirectory, getHomeDirectory(throwIfNull:true));
-      expect(p.atKeysFilePath, getDefaultAtKeysFilePath(p.homeDirectory, p.clientAtSign));
+      expect(p.homeDirectory, getHomeDirectory(throwIfNull: true));
+      expect(p.atKeysFilePath,
+          getDefaultAtKeysFilePath(p.homeDirectory, p.clientAtSign));
       expect(p.sendSshPublicKey, 'false');
       expect(p.localSshOptions, []);
       expect(p.rsa, false);
@@ -59,19 +60,25 @@ void main() {
       args.addAll(['-t', '@bob']);
       args.addAll(['-h', 'host.subdomain.test']);
 
-
       args.addAll([
-        '--device','ancient_pc',
-        '--port','56789',
-        '--local-port','98765',
-        '--key-file','/tmp/temp_keys.json',
-        '--ssh-public-key','sekrit.pub',
-        '--local-ssh-options','--arg 2 --arg 4 foo bar -x',
-        '--remote-user-name','gary',
+        '--device',
+        'ancient_pc',
+        '--port',
+        '56789',
+        '--local-port',
+        '98765',
+        '--key-file',
+        '/tmp/temp_keys.json',
+        '--ssh-public-key',
+        'sekrit.pub',
+        '--local-ssh-options',
+        '--arg 2 --arg 4 foo bar -x',
+        '--remote-user-name',
+        'gary',
         '-v',
         '-r'
       ]);
-      var p = SSHNP.parseSSHNPParams(args);
+      var p = SSHNPParams.fromPartial(SSHNPPartialParams.fromArgs(args));
       expect(p.clientAtSign, '@alice');
       expect(p.sshnpdAtSign, '@bob');
       expect(p.host, 'host.subdomain.test');
@@ -80,9 +87,10 @@ void main() {
       expect(p.port, '56789');
       expect(p.localPort, '98765');
       expect(p.username, getUserName(throwIfNull: true));
-      expect(p.homeDirectory, getHomeDirectory(throwIfNull:true));
+      expect(p.homeDirectory, getHomeDirectory(throwIfNull: true));
       expect(p.atKeysFilePath, '/tmp/temp_keys.json');
-      expect(p.sendSshPublicKey, '${getDefaultSshDirectory(p.homeDirectory)}sekrit.pub');
+      expect(p.sendSshPublicKey,
+          '${getDefaultSshDirectory(p.homeDirectory)}sekrit.pub');
       expect(p.localSshOptions, ['--arg 2 --arg 4 foo bar -x']);
       expect(p.rsa, true);
       expect(p.verbose, true);
